refactor(Server): Move the server in the Engine directory
It involves moving all the server folder, but also changing all the paths from cmakes, headers or classic c++ files
