+++ Engine/Server/CMakeLists.txt
@@ -2,10 +2,12 @@ cmake_minimum_required(VERSION 3.5)
 
 set(EXE r-type_server)
 
-set(EXE_SRC_PATH "src")
+set(EXE_SRC_PATH "Src")
 set(EXE_INCLUDE_PATH
-    "include"
-    "src/GameEngine"
+    "Include"
+    "Src/GameEngine"
+    "Src/Network"
+    "Src/GameEngine/ServerSceneManager"
 )
 
 set(CMAKE_CXX_STANDARD 20)
@@ -14,28 +16,28 @@ set(SRC_EXE
     ./${EXE_SRC_PATH}/main.cpp
     ./${EXE_SRC_PATH}/GameEngine/Application/Application.cpp
     ./${EXE_SRC_PATH}/GameEngine/ServerSceneManager/ServerSceneManager.cpp
-    ../SceneManager/ASceneManager.cpp
+    ../Shared/SceneManager/ASceneManager.cpp
 )
 
 include_directories(
-    ../Ecs/src/
-    ../Ecs/include/
-    ./src/Network/
-    ../include/
-    ../include/interfaces/
-    ./src/GameEngine/Application/
-    ./src/Errors/
-    ../SceneManager/
+        src/Network
+        src/GameEngine/Errors
+        src/GameEngine/Application
+        src/GameEngine/ServerSceneManager
+        ../Shared/Ecs/Src
+        ../Shared/Interface
+        ../Shared/SceneManager
+        ../Shared/Network/Packet
 )
 
-add_subdirectory(src/Network)
+add_subdirectory(Src/Network)
 
 add_executable(${EXE} ${SRC_EXE})
 
 target_include_directories(${EXE} PUBLIC ${EXE_INCLUDE_PATH})
 
-find_package(asio CONFIG REQUIRED)
-find_package(jsoncpp CONFIG REQUIRED)
+find_package(asio REQUIRED)
+find_package(jsoncpp REQUIRED)
 
 target_link_libraries(
     ${EXE}
